requirements.txt
tweepy>=4.14,<5
python-dotenv>=1.0
